# mobile/openapi.yaml
openapi: 3.0.3
info:
  title: ViVu App API
  version: 1.0.0
servers:
  - url: http://localhost:8080
    description: Local Development
  - url: https://api.vivuapp.vn
    description: Production
...
paths:
...
  /api/v1/conversations/create:
    post:
      tags: [Conversations]
      security:
        - bearerAuth: []
      summary: Create private conversation
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ConversationPrivateRequest"
      responses:
        "201":
          description: Conversation id
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiResponseString"
  /api/v1/conversations/private:
    get:
      tags: [Conversations]
      security:
        - bearerAuth: []
      summary: List private conversations
      parameters:
        - in: query
          name: page
          schema:
            type: integer
            default: 0
        - in: query
          name: size
          schema:
            type: integer
            default: 10
      responses:
        "200":
          description: Slice of conversations
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiResponseConversationSlice"
  /api/v1/conversations/channels:
    get:
      tags: [Conversations]
      security:
        - bearerAuth: []
      summary: List channel conversations
      parameters:
        - in: query
          name: page
          schema:
            type: integer
            default: 0
        - in: query
          name: size
          schema:
            type: integer
            default: 10
      responses:
        "200":
          description: Slice of conversations
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiResponseConversationSlice"
  /api/v1/conversations/delete-conversation/{conversationId}:
    delete:
      tags: [Conversations]
      security:
        - bearerAuth: []
      summary: Leave or delete conversation
      parameters:
        - in: path
          name: conversationId
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Deleted
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiResponseString"
  /api/v1/conversations/{conversationId}/media:
    get:
      tags: [Conversations]
      security:
        - bearerAuth: []
      summary: Get conversation media
      parameters:
        - in: path
          name: conversationId
          required: true
          schema:
            type: string
        - in: query
          name: fileType
          schema:
            $ref: "#/components/schemas/FileType"
      responses:
        "200":
          description: Media attachments
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiResponseMediaAttachmentList"
  /api/v1/channels:
    post:
      tags: [Channels]
      security:
        - bearerAuth: []
      summary: Create channel
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema:
              type: object
              required: [name]
              properties:
                name:
                  type: string
                description:
                  type: string
                privacy:
                  $ref: "#/components/schemas/ChannelPrivacy"
                channelType:
                  $ref: "#/components/schemas/ChannelType"
                participantIds:
                  type: string
                  description: JSON array string, e.g. [1,2,3]
                avatar:
                  type: string
                  format: binary
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiResponseChannelResponse"
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
    ApiResponseBase:
      type: object
      properties:
        code:
          type: integer
        message:
          type: string
        result:
          nullable: true
...
    ApiResponseString:
      allOf:
        - $ref: "#/components/schemas/ApiResponseBase"
        - type: object
          properties:
            result:
              type: string
...
    ApiResponseConversationSlice:
      allOf:
        - $ref: "#/components/schemas/ApiResponseBase"
        - type: object
          properties:
            result:
              $ref: "#/components/schemas/SliceConversationResponse"
    ApiResponseMediaAttachmentList:
      allOf:
        - $ref: "#/components/schemas/ApiResponseBase"
        - type: object
          properties:
            result:
              type: array
              items:
                $ref: "#/components/schemas/MediaAttachmentResponse"
    ApiResponseChannelResponse:
      allOf:
        - $ref: "#/components/schemas/ApiResponseBase"
        - type: object
          properties:
            result:
              $ref: "#/components/schemas/ChannelResponse"
...
    ConversationPrivateRequest:
      type: object
      properties:
        receiverId:
          type: integer
          format: int64
...
    ChannelResponse:
      type: object
      properties:
        id:
          type: integer
          format: int64
        name:
          type: string
        description:
          type: string
        privacy:
          $ref: "#/components/schemas/ChannelPrivacy"
        channelType:
          $ref: "#/components/schemas/ChannelType"
        avatarUrl:
          type: string
        conversationId:
          type: string
        creatorId:
          type: integer
          format: int64
        memberCount:
          type: integer
        createdAt:
          type: string
          format: date-time
...
    ConversationResponse:
      type: object
      properties:
        id:
          type: string
        name:
          type: string
        avatarUrl:
          type: string
        unreadCount:
          type: integer
          format: int64
        lastMessage:
          type: string
        lastMessageTime:
          type: string
          format: date-time
...
    MediaAttachmentResponse:
      type: object
      properties:
        id:
          type: integer
          format: int64
        messageId:
          type: integer
          format: int64
        fileId:
          type: integer
          format: int64
        fileUrl:
          type: string
        fileType:
          $ref: "#/components/schemas/FileType"
        fileSize:
          type: integer
          format: int64
        uploadedAt:
          type: string
          format: date-time
...
    Pageable:
      type: object
      properties:
        pageNumber:
          type: integer
        pageSize:
          type: integer
        sort:
          $ref: "#/components/schemas/Sort"
        offset:
          type: integer
        paged:
          type: boolean
        unpaged:
          type: boolean
    Sort:
      type: object
      properties:
        sorted:
          type: boolean
        unsorted:
          type: boolean
        empty:
          type: boolean
...
    SliceConversationResponse:
      type: object
      properties:
        content:
          type: array
          items:
            $ref: "#/components/schemas/ConversationResponse"
        pageable:
          $ref: "#/components/schemas/Pageable"
        size:
          type: integer
        number:
          type: integer
        sort:
          $ref: "#/components/schemas/Sort"
        numberOfElements:
          type: integer
        first:
          type: boolean
        last:
          type: boolean
        empty:
          type: boolean
...
    ChannelPrivacy:
      type: string
      enum: [PUBLIC, PRIVATE]
    ChannelType:
      type: string
      enum: [TEXT, VOICE, VIDEO, ANNOUNCEMENT]
...
    FileType:
      type: string
      enum: [IMAGE, VIDEO, AUDIO, DOCUMENT]